+++ zcash_primitives/src/legacy.rs
@@ -6,6 +6,11 @@ use std::fmt;
 use std::io::{self, Read, Write};
 use std::ops::Shl;
 
+#[cfg(feature = "transparent-inputs")]
+use hdwallet::KeyIndex;
+
+use subtle::{Choice, ConstantTimeEq};
+
 use zcash_encoding::Vector;
 
 #[cfg(feature = "transparent-inputs")]
@@ -402,6 +407,61 @@ impl TransparentAddress {
     }
 }
 
+/// A child index for a derived transparent address.
+///
+/// Only NON-hardened derivation is supported.
+#[derive(Clone, Copy, Debug, PartialEq, Eq)]
+pub struct NonHardenedChildIndex(u32);
+
+impl ConstantTimeEq for NonHardenedChildIndex {
+    fn ct_eq(&self, other: &Self) -> Choice {
+        self.0.ct_eq(&other.0)
+    }
+}
+
+impl NonHardenedChildIndex {
+    pub const ZERO: NonHardenedChildIndex = NonHardenedChildIndex(0);
+
+    /// Parses the given ZIP 32 child index.
+    ///
+    /// Returns `None` if the hardened bit is set.
+    pub fn from_index(i: u32) -> Option<Self> {
+        if i < (1 << 31) {
+            Some(NonHardenedChildIndex(i))
+        } else {
+            None
+        }
+    }
+
+    /// Returns the index as a 32-bit integer.
+    pub fn index(&self) -> u32 {
+        self.0
+    }
+
+    pub fn next(&self) -> Option<Self> {
+        Self::from_index(self.0 + 1)
+    }
+}
+
+#[cfg(feature = "transparent-inputs")]
+impl TryFrom<KeyIndex> for NonHardenedChildIndex {
+    type Error = ();
+
+    fn try_from(value: KeyIndex) -> Result<Self, Self::Error> {
+        match value {
+            KeyIndex::Normal(i) => NonHardenedChildIndex::from_index(i).ok_or(()),
+            KeyIndex::Hardened(_) => Err(()),
+        }
+    }
+}
+
+#[cfg(feature = "transparent-inputs")]
+impl From<NonHardenedChildIndex> for KeyIndex {
+    fn from(value: NonHardenedChildIndex) -> Self {
+        Self::Normal(value.index())
+    }
+}
+
 #[cfg(any(test, feature = "test-dependencies"))]
 pub mod testing {
     use proptest::prelude::{any, prop_compose};
@@ -417,7 +477,7 @@ pub mod testing {
 
 #[cfg(test)]
 mod tests {
-    use super::{OpCode, Script, TransparentAddress};
+    use super::{NonHardenedChildIndex, OpCode, Script, TransparentAddress};
 
     #[test]
     fn script_opcode() {
@@ -483,5 +543,22 @@ mod tests {
             ]
         );
         assert_eq!(addr.script().address(), Some(addr));
+    }
+
+    #[test]
+    fn nonhardened_indexes_accepted() {
+        assert_eq!(0, NonHardenedChildIndex::from_index(0).unwrap().index());
+        assert_eq!(
+            0x7fffffff,
+            NonHardenedChildIndex::from_index(0x7fffffff)
+                .unwrap()
+                .index()
+        );
+    }
+
+    #[test]
+    fn hardened_indexes_rejected() {
+        assert!(NonHardenedChildIndex::from_index(0x80000000).is_none());
+        assert!(NonHardenedChildIndex::from_index(0xffffffff).is_none());
     }
 }