+++ zcash_primitives/src/legacy/keys.rs
@@ -6,14 +6,11 @@ use hdwallet::{
 };
 use secp256k1::PublicKey;
 use sha2::{Digest, Sha256};
-use subtle::{Choice, ConstantTimeEq};
 use zcash_spec::PrfExpand;
 
 use crate::{consensus, zip32::AccountId};
 
-use super::TransparentAddress;
-
-const MAX_TRANSPARENT_CHILD_INDEX: u32 = 0x7FFFFFFF;
+use super::{NonHardenedChildIndex, TransparentAddress};
 
 /// A [BIP44] private key at the account path level `m/44'/<coin_type>'/<account>'`.
 ///
@@ -51,11 +48,11 @@ impl AccountPrivKey {
     /// `m/44'/<coin_type>'/<account>'/0/<child_index>`.
     pub fn derive_external_secret_key(
         &self,
-        child_index: u32,
+        child_index: NonHardenedChildIndex,
     ) -> Result<secp256k1::SecretKey, hdwallet::error::Error> {
         self.0
             .derive_private_key(KeyIndex::Normal(0))?
-            .derive_private_key(KeyIndex::Normal(child_index))
+            .derive_private_key(child_index.into())
             .map(|k| k.private_key)
     }
 
@@ -187,30 +184,31 @@ pub trait IncomingViewingKey: private::SealedChangeLevelKey + std::marker::Sized
     #[allow(deprecated)]
     fn derive_address(
         &self,
-        child_index: u32,
+        child_index: NonHardenedChildIndex,
     ) -> Result<TransparentAddress, hdwallet::error::Error> {
         let child_key = self
             .extended_pubkey()
-            .derive_public_key(KeyIndex::Normal(child_index))?;
+            .derive_public_key(child_index.into())?;
         Ok(pubkey_to_address(&child_key.public_key))
     }
 
     /// Searches the space of child indexes for an index that will
     /// generate a valid transparent address, and returns the resulting
     /// address and the index at which it was generated.
-    fn default_address(&self) -> (TransparentAddress, u32) {
-        let mut child_index = 0;
-        while child_index <= MAX_TRANSPARENT_CHILD_INDEX {
+    fn default_address(&self) -> (TransparentAddress, NonHardenedChildIndex) {
+        let mut child_index = NonHardenedChildIndex::ZERO;
+        loop {
             match self.derive_address(child_index) {
                 Ok(addr) => {
                     return (addr, child_index);
                 }
                 Err(_) => {
-                    child_index += 1;
+                    child_index = child_index.next().unwrap_or_else(|| {
+                        panic!("Exhausted child index space attempting to find a default address.");
+                    });
                 }
             }
         }
-        panic!("Exhausted child index space attempting to find a default address.");
     }
 
     fn serialize(&self) -> Vec<u8> {
@@ -292,39 +290,9 @@ impl ExternalOvk {
     }
 }
 
-/// A child index for a derived transparent address.
-///
-/// Only NON-hardened derivation is supported.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub struct NonHardenedChildIndex(u32);
-
-impl ConstantTimeEq for NonHardenedChildIndex {
-    fn ct_eq(&self, other: &Self) -> Choice {
-        self.0.ct_eq(&other.0)
-    }
-}
-
-impl NonHardenedChildIndex {
-    /// Parses the given ZIP 32 child index.
-    ///
-    /// Returns `None` if the hardened bit is set.
-    pub fn from_index(i: u32) -> Option<Self> {
-        if i < (1 << 31) {
-            Some(NonHardenedChildIndex(i))
-        } else {
-            None
-        }
-    }
-
-    /// Returns the index as a 32-bit integer.
-    pub fn index(&self) -> u32 {
-        self.0
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use super::{AccountPubKey, NonHardenedChildIndex};
+    use super::AccountPubKey;
 
     #[test]
     fn check_ovk_test_vectors() {
@@ -570,22 +538,5 @@ mod tests {
             assert_eq!(tv.internal_ovk, internal.as_bytes());
             assert_eq!(tv.external_ovk, external.as_bytes());
         }
-    }
-
-    #[test]
-    fn nonhardened_indexes_accepted() {
-        assert_eq!(0, NonHardenedChildIndex::from_index(0).unwrap().index());
-        assert_eq!(
-            0x7fffffff,
-            NonHardenedChildIndex::from_index(0x7fffffff)
-                .unwrap()
-                .index()
-        );
-    }
-
-    #[test]
-    fn hardened_indexes_rejected() {
-        assert!(NonHardenedChildIndex::from_index(0x80000000).is_none());
-        assert!(NonHardenedChildIndex::from_index(0xffffffff).is_none());
     }
 }